Fix typo in get_random_int parameters that took floats instead of ints

+++ include/random.h
@@ -8,7 +8,7 @@
 #ifndef RANDOM_H_
     #define RANDOM_H_
 
-    int get_random_int(float const min, float const max);
+    int get_random_int(int const min, int const max);
     float get_random_float(float const min, float const max);
     int get_random_choice_int(int choice1, int choice2);
     sfColor get_random_color(unsigned int opacity);
